Initial setup for encryption of /home

+++ install_artix.sh
@@ -1193,15 +1193,23 @@ EOM
 
   SCRIPT_06_FORMAT_AND_ENCRYPT_PARTITIONS() {
     mkfs.vfat -F32 -n "$BOOT_label" "$DRIVE_path_boot" 
-    if [[ "$HOME_partition" == "true" ]]; then
+    if [[ "$HOME_partition" == "true" ]] && ! [[ "$ENCRYPTION_partitions" == "true" ]]; then
       mkfs.btrfs -f -L "$HOME_label" "$DRIVE_path_home"
     fi 
     if [[ "$FILESYSTEM_primary_btrfs" == "true" ]]; then
       if [[ "$ENCRYPTION_partitions" == "true" ]]; then
-        echo "$ENCRYPTION_passwd" | cryptsetup luksFormat --batch-mode --type luks2 --pbkdf pbkdf2 --cipher aes-xts-plain64 --key-size 512 --hash sha512 --use-random "$DRIVE_path_primary" # GRUB currently lacks support for ARGON2d
-        echo "$ENCRYPTION_passwd" | cryptsetup open --allow-discards --perf-no_read_workqueue --size 4196 --persistent "$DRIVE_path_primary" cryptroot
-        mkfs.btrfs -f -L "$PRIMARY_label" /dev/mapper/cryptroot
-        MOUNTPOINT="/dev/mapper/cryptroot"
+        if [[ "$HOME_partition" == "true" ]]; then
+          echo "$ENCRYPTION_passwd" | cryptsetup luksFornat --batch-mode --type luks2 --cipher aes-xts_plain64 --key-size 512 --hash sha512 --use-random "$DRIVE_path_home"
+          echo "$ENCRYPTION_passwd" | cryptsetup open --allow-discards --perf-no_read_workqueue --size 4196 --persistent "$DRIVE_path_home" cryptroot
+          mkfs.btrfs -f -L "$HOME_label" "$DRIVE_path_home"
+          mkfs.btrfs -f -L "$PRIMARY_label" "$DRIVE_path_primary"
+          MOUNTPOINT="/dev/mapper/cryptroot"
+        else
+          echo "$ENCRYPTION_passwd" | cryptsetup luksFormat --batch-mode --type luks2 --pbkdf pbkdf2 --cipher aes-xts-plain64 --key-size 512 --hash sha512 --use-random "$DRIVE_path_primary" # GRUB currently lacks support for ARGON2d
+          echo "$ENCRYPTION_passwd" | cryptsetup open --allow-discards --perf-no_read_workqueue --size 4196 --persistent "$DRIVE_path_primary" cryptroot
+          mkfs.btrfs -f -L "$PRIMARY_label" /dev/mapper/cryptroot
+          MOUNTPOINT="/dev/mapper/cryptroot"
+        fi
       else
         mkfs.btrfs -f -L "$PRIMARY_label" "$DRIVE_path_primary"
         MOUNTPOINT="$DRIVE_path_primary"
